requirements.txt
pycryptodome>=3.20